Use pydantic and split up train script
